Updated petsc-rs to use num-complex `Complex` type
Right now, petsc-sys still uses the `__BindgenComplex` type but I impl'ed an into that transmutes it to `Complex`. Then on the petsc-rs side, we call `.into()` so that the compiler sees that all the types line up. if we have a pointer, we can trivially cast it i.e. `as *mut _`.

I also updated all of the examples (including doc-string examples) to support a complex petsc install. This makes things look a lot more ugly, but I think it's fine.

Also added support for a release install (it uses the arch value in `PETSC_ARCH_RELEASE` instead of `PETSC_ARCH` when you build the rust with release mode).

I have tested with a petsc install that uses real f64 + i32 and one that uses complex f32 + i64.