Made MatShell its own struct
This was done because we want to let the user have mutable access to some user-defined `mat_data`. This can't be done with closures alone because rust would think there are multiple mutable references. So we created a method `get_mat_data_mut` which uses a generic over the mat data. To not have a generic on `Mat`, we created a new struct `MatShell` to use the generic.

Also removed many uses of `Pin::get_unchecked_mut()` because they were unneeded (because `Box` `impl`s `Unpin`).

Added `shell_set_operation_mvi` method